Update tests for new gold alignment
* Move relevant tests from `test_align.py` into `test_gold.py`
* Remove `test_align.py`
